MAINT: Refactor
