+++ .coveragerc
@@ -14,3 +14,5 @@ omit =
 exclude_lines =
     pragma: no cover
     if __name__ == .__main__.:
+    @abstractmethod
+    @abstractclassmethod
